// src/app/api/generated/model/espacio-page-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EspacioResponse } from './espacio-response';


/**
 * Pagina de espacios de la sede, segun el filtro de estado pedido
 */
export interface EspacioPageResponse { 
    /**
     * Espacios de esta pagina
     */
    content?: Array<EspacioResponse>;
    /**
     * Numero de pagina devuelta, base cero
     */
    page?: number;
    /**
     * Cantidad de elementos por pagina
     */
    size?: number;
    /**
     * Cantidad total de espacios que cumplen el filtro
     */
    totalElements?: number;
    /**
     * Cantidad total de paginas disponibles
     */
    totalPages?: number;
}